fix: delete app menu item is only visible to app creators and admins, only creator or admins can delete apps

+++ src/lib/components/workspace/Apps/AppMenu.svelte
@@ -175,17 +175,19 @@
 				</DropdownMenu.Item>
 			{/if}
 
-			<hr class="border-gray-50/30 dark:border-gray-800/30 my-1" />
+			{#if $currentUser?.role === 'admin' || app?.user_id === $currentUser?.id}
+				<hr class="border-gray-50/30 dark:border-gray-800/30 my-1" />
 
-			<DropdownMenu.Item
-				class="flex  gap-2  items-center px-3 py-1.5 text-sm  cursor-pointer hover:bg-gray-50 dark:hover:bg-gray-800 rounded-xl"
-				on:click={() => {
-					deleteHandler();
-				}}
-			>
-				<GarbageBin />
-				<div class="flex items-center">{$i18n.t('Delete')}</div>
-			</DropdownMenu.Item>
+				<DropdownMenu.Item
+					class="flex  gap-2  items-center px-3 py-1.5 text-sm  cursor-pointer hover:bg-gray-50 dark:hover:bg-gray-800 rounded-xl"
+					on:click={() => {
+						deleteHandler();
+					}}
+				>
+					<GarbageBin />
+					<div class="flex items-center">{$i18n.t('Delete')}</div>
+				</DropdownMenu.Item>
+			{/if}
 		</DropdownMenu.Content>
 	</div>
 </Dropdown>
